adds additional chapter end section tests
